Migrate the cloudantplus node into this module
