--- .cfignore
@@ -1,2 +1,0 @@
-launchConfigurations/
-.git/
