+++ cloudant.html
@@ -1,5 +1,5 @@
 <!--
-  Copyright 2014,2016 IBM Corp.
+  Copyright 2014, 2020 IBM Corp.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -114,7 +114,9 @@
         <label for="node-input-search"><i class="fa fa-search"></i> Search by</label>
         <select id="node-input-search" style="width:68%;">
             <option value="_id_">_id</option>
+            <option value="_query_">query</option>
             <option value="_idx_">search index</option>
+            <option value="_view_">view</option>
             <option value="_all_">all documents</option>
         </select>
     </div>
@@ -122,7 +124,7 @@
         <label>&nbsp;</label>
         <input type="text" id="node-input-design" style="width:30%" placeholder="design document">
         /
-        <input type="text" id="node-input-index" style="width:30%" placeholder="index name">
+        <input type="text" id="node-input-index" style="width:30%" placeholder="index/view name">
     </div>
     <div class="form-row">
         <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
@@ -143,7 +145,7 @@
             operation: { value:"insert" }
         },
         inputs: 1,
-        outputs: 0,
+        outputs: 1,
         icon: "cloudant.png",
         align: "right",
         label: label,
@@ -214,7 +216,7 @@
         searchBySelect.change(function() {
             var searchBy = searchBySelect.val();
 
-            if (searchBy === "_idx_") {
+            if (searchBy === "_idx_" || searchBy === "_view_") {
                 node._def.defaults.design.required = true;
                 node._def.defaults.index.required  = true;
                 searchForm.show();
@@ -243,6 +245,7 @@
 <script type="text/x-red" data-help-name="cloudant out">
    <p>
        A simple Cloudant output node. Stores <b>msg</b> in a chosen database.
+       If <b>msg</b> is an array, all elements will be processed using the bulk operations API.
     </p>
    <p>
        Cloudant generates a random id for the <code>_id</code> property if one
@@ -253,7 +256,7 @@
    <p>
        If you want to <b>update</b> an existing document you must specify the
        latest values for <code>_id</code> and <code>_rev</code> in your object.
-       If they don't match the current values stored in the database you will
+       If they don&apos;t match the current values stored in the database you will
        receive a <b>conflict error</b> message, so make sure you are using the
        most updated information for the document.
     </p>
@@ -261,15 +264,15 @@
         It is also possible to <b>delete</b> documents from the database by
         providing values for <code>_id</code> and <code>_rev</code> and selecting
         the <b>remove</b> option for the node. You can pass these values in the
-        <code>msg</code> object itself or as an object in the
+        <code>msg</code> object itself or as an object or array of objects in the
         <code>msg.payload</code>.
     </p>
     <p>
         The <b>database name</b> must follow these rules:
         <ul>
             <li>No spaces</li>
-            <li>All letters in small caps</li>
-            <li>The first character can't be <code>_</code></li>
+            <li>All letters in lower case</li>
+            <li>The first character can&apos;t be <code>_</code></li>
         </ul>
     </p>
     <p>
@@ -285,15 +288,22 @@
         A node for searching documents in a Cloudant database.
     </p>
     <p>
-        Searching for documents can be done in three modes: directly by using the
-        document's <b>_id</b>, by using an existing <a
-        href="https://cloudant.com/for-developers/search/" target="_blank">Search
-        Index</a> or retrieving <b>all documents</b> stored in your database.
+        Searching for documents can be done in five modes: directly by using the
+        document&apos;s <b>_id</b>, by using a <a href="https://console.bluemix.net/docs/services/Cloudant/api/cloudant_query.html#finding-documents-using-an-index"  target="_blank"
+        declarative JSON query</a>, by using an existing <a
+        href="https://console.bluemix.net/docs/services/Cloudant/api/search.html#search" target="_blank">Search
+        Index</a>, by using a <a href="https://console.bluemix.net/docs/services/Cloudant/api/using_views.html#using-views" target="_blank">
+        view</a> or retrieving <b>all documents</b> stored in your database.
     </p>
     <p>
-        When querying using the <b>_id</b> option, the value for the document's
+        When querying using the <b>_id</b> option, the value for the document&apos;s
         <code>_id</code> should be passed in the <code>msg.payload</code> as a
         string.
+    </p>
+    <p>
+        When querying using the <b>query</b> option, the JSON object with
+        <a href="https://console.bluemix.net/docs/services/Cloudant/api/cloudant_query.html#finding-documents-using-an-index" target="_blank">
+        declarative query syntax</a> should be passed in the <code>msg.payload</code>.
     </p>
     <p>
         To use an existing <b>Search Index</b> stored on the desired database,
@@ -318,6 +328,13 @@
         official Cloudant documentation</a>.
     </p>
     <p>
+      When querying using a <b>View</b> you can pass the view parameters as an
+      object in <code>msg.payload</code>. This object should contain key/value
+      pairs as defined in the Query string section in the
+      <a href="https://console.bluemix.net/docs/services/Cloudant/api/using_views.html#using-views">
+      Cloudant documentation</a>
+    </p>
+    <p>
         The last method to retrieve documents is to simply get all of them by
         selecting the option <b>all documents</b>.
     </p>
@@ -325,8 +342,8 @@
         The <b>database name</b> must follow these rules:
         <ul>
             <li>No spaces</li>
-            <li>All letters in small caps</li>
-            <li>The first character can't be <code>_</code></li>
+            <li>All letters in lower case</li>
+            <li>The first character can&apos;t be <code>_</code></li>
         </ul>
     </p>
 </script>
