--- project.json
@@ -1,1 +1,0 @@
-{"Name":"laoqui2 | node-red-node-cf-cloudant"}
